# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trackventory API",
        "description": "API documentation for Trackventory - Inventory Tracking System",
        "contact": {
            "name": "Trackventory Support",
            "email": "support@trackventory.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/borrow-statuses": {
            "get": {
                "tags": [
                    "Borrow Statuses"
                ],
                "summary": "Get all borrow statuses",
                "description": "Retrieve a list of all borrow statuses",
                "operationId": "c69e85f4e2e6f6a79c6912c5f0a4a16b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow statuses retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BorrowStatus"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Borrow Statuses"
                ],
                "summary": "Create a new borrow status",
                "description": "Create a new borrow status with status name and badge color",
                "operationId": "433ae6d252c1f8d97ed5965880c48fb0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_name",
                                    "badge_color"
                                ],
                                "properties": {
                                    "status_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Borrowed"
                                    },
                                    "badge_color": {
                                        "type": "string",
                                        "maxLength": 7,
                                        "example": "#007BFF"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Borrow status created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow status created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BorrowStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/borrow-statuses/{id}": {
            "get": {
                "tags": [
                    "Borrow Statuses"
                ],
                "summary": "Get a specific borrow status",
                "description": "Retrieve a specific borrow status by ID",
                "operationId": "6e963b111a99deee2d77789c6a0a564d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Borrow status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow status retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BorrowStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Borrow status not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Borrow Statuses"
                ],
                "summary": "Update a borrow status",
                "description": "Update an existing borrow status",
                "operationId": "acf1675a0f78b07d70b74de7c88858f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Borrow status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Returned"
                                    },
                                    "badge_color": {
                                        "type": "string",
                                        "maxLength": 7,
                                        "example": "#28A745"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Borrow status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Borrow status updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BorrowStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Borrow status not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Borrow Statuses"
                ],
                "summary": "Delete a borrow status",
                "description": "Delete an existing borrow status",
                "operationId": "1827f5612fc2872833af29bbef2cc813",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Borrow status ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Borrow status deleted successfully"
                    },
                    "404": {
                        "description": "Borrow status not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BorrowStatus": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status_name": {
                        "type": "string",
                        "example": "Borrowed"
                    },
                    "badge_color": {
                        "type": "string",
                        "example": "#007BFF"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}